Updates to deployment
- helloworld gets an extra sidecar container
- This is for testing M:N service:container mappings
- Remove fluxy deployment yamels
