--- deploy/fluxy-deployment.yaml
@@ -1,36 +1,0 @@
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  name: fluxy
-spec:
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        name: fluxy
-    spec:
-      volumes:
-      - name: key
-        secret:
-          secretName: fluxy-repo-key
-      containers:
-      - name: fluxy
-        image: weaveworks/fluxy
-        imagePullPolicy: Never # must build manually
-        ports:
-        - containerPort: 3030
-        volumeMounts:
-        - name: key
-          mountPath: /var/run/secrets/fluxy/key
-          readOnly: true
-        args:
-        - /home/flux/fluxd
-        - --kubernetes-kubectl=/home/flux/kubectl
-        - --kubernetes-host=https://kubernetes
-        - --kubernetes-certificate-authority=/var/run/secrets/kubernetes.io/serviceaccount/ca.crt
-        - --kubernetes-bearer-token-file=/var/run/secrets/kubernetes.io/serviceaccount/token
-        - --database-driver=ql
-        - --database-source=file://history.db
-        - --repo-url=git@github.com:squaremo/fluxy-testdata
-        - --repo-key=/var/run/secrets/fluxy/key/id-rsa
-        - --repo-path=testdata
