--- testdata/Dockerfile
@@ -1,4 +1,0 @@
-FROM scratch
-ENTRYPOINT ["/helloworld"]
-EXPOSE 80
-COPY ./helloworld-linux-amd64 /helloworld
