+++ testdata/helloworld-deploy.yaml
@@ -16,3 +16,9 @@ spec:
         - -msg=Ahoy
         ports:
         - containerPort: 80
+      - name: sidecar
+        image: quay.io/weaveworks/sidecar:master-a000001
+        args:
+        - -addr=:8080
+        ports:
+        - containerPort: 8080
