+++ testdata/helloworld/Dockerfile
@@ -1,0 +1,4 @@
+FROM scratch
+ENTRYPOINT ["/helloworld"]
+EXPOSE 80
+COPY ./helloworld-linux-amd64 /helloworld
