+++ testdata/sidecar/Dockerfile
@@ -1,0 +1,4 @@
+FROM scratch
+ENTRYPOINT ["/sidecar"]
+EXPOSE 8080
+COPY ./sidecar-linux-amd64 /sidecar
